move tests into own folder
