+++ ctts/tests/test_prediction_api.py
@@ -1,7 +1,7 @@
 from fastapi.testclient import TestClient
 
-from .api import app
-from .constants import API_PREFIX
+from ..api import app
+from ..constants import API_PREFIX
 
 client = TestClient(app)
